vagrant: allow network communications between all vms
Before this commit, the vm can only communicate with the vlan on the
same /24 network range. It's ok for most of the use case but doesn't
allow to test staging webapp log ingestion for example.
All the vms are now on the same /16 network. It implies to remove all
the previous vagrant networks. The basic use cases will still work if
not.